refactor: project structure
edit plan Schema, style: quotes
