+++ miamoremustachio/ToDo/modules/models/plan-model.js
@@ -3,7 +3,7 @@ const mongoose = require('mongoose');
 const planSchema = new mongoose.Schema({
   title: { type: String, required: true },
   tasks: [{ type: 'ObjectId', ref: 'Task' }],
-  user: { type: 'ObjectId', ref: 'User' },
+  user: { type: 'ObjectId', ref: 'User', immutable: true },
 },
 {
   timestamps: true,
